requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
awsebcli==3.20.3
beautifulsoup4==4.9.3
botocore==1.23.54
cement==2.8.2
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
charset-normalizer==2.0.4
colorama
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
django_heroku
Django==3.2.5
django-admin-honeypot==1.1.0
django-allauth==0.50.0
django-cors-headers==3.13.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-filter==21.1
django-formtools==2.3
django-heroku==0.3.1
django-jet-reboot==1.3.1
django-multiselectfield==0.1.12
django-otp==1.0.6
django-paypal==2.0
django-phonenumber-field==5.2.0
django-robots==5.0
django-two-factor-auth==1.13.1
djangorestframework==3.13.1
future==0.16.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.11.3
incremental==21.3.0
jmespath==0.10.0
Markdown==3.3.4
markdown2==2.4.0
msgpack==1.0.3
mysqlclient==2.0.3
numpy==1.22.3
oauthlib==3.2.0
pathspec==0.9.0
pep8==1.7.1
phonenumbers==8.12.29
Pillow==8.3.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.1
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
qrcode==6.1
requests==2.26.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
service-identity==21.1.0
sgmllib3k==1.0.0
six
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.74.0
termcolor==1.1.0
Twisted==22.4.0
txaio==22.2.1
typing-extensions==3.10.0.0
tzdata==2022.1
urllib3==1.26.6
wcwidth==0.1.9
whitenoise==5.2.0
zipp==3.8.0
zope.interface==5.4.0
